pyproject.toml
[tool.poetry]
name = "aurora"
version = "0.0.1"
description = ""
authors = ["ZharovVsevolod <wsewolod@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gigachain-community = "^0.2.12"
python-dotenv = "^1.0.1"
faiss-cpu = "^1.8.0.post1"
pandas = "^2.2.3"
python-telegram-bot = {version = "^21.5", optional = true}
fastapi = {extras = ["standard"], version = "^0.115.0", optional = true}
uvicorn = {extras = ["standard"], version = "^0.30.6", optional = true}

[tool.poetry.extras]
telegram = ["python-telegram-bot"]
api = ["fastapi", "uvicorn"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
